// UI/src/app/core/models/animal.enums.ts
export enum TipoAnimal {
  Hembra = 1,
  Macho = 2,
}

export enum PropositoAnimal {
  Carne = 1,
  Leche = 2,
  Mixto = 3,
}

export enum EstadoHembra {
  Cria = 1,
  Recria = 2,
  Novilla = 3,
  Proxima = 4,
  Parida = 5,
  Escotera = 6,
}

export enum EstadoMacho {
  Cria = 1,
  Recria = 2,
  Torete = 3,
  Toro = 4,
}